backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
httpx
python-dotenv
sqlalchemy
pypdf2
python-docx
markdown
sentence-transformers
numpy
pytesseract
pdf2image
openai>=1.0.0 